before-lunch

+++ controllers/property.js
@@ -5,6 +5,7 @@ const forRent = require("../propertiesModule/forRentHsA");
 const forSale = require("../propertiesModule/forSaleHsA");
 const landsAndPlots = require("../propertiesModule/landsAndPlots");
 const forRentOfficeShops = require("../propertiesModule/forRentSsO");
+const User = require("../modules/user");
 
 const clearImage = (filePath) => {
   filePath = path.join(__dirname, "..", filePath);
@@ -12,7 +13,6 @@ const clearImage = (filePath) => {
 };
 
 exports.postRent = async (req, res, next) => {
-  // console.log("Ready")
   const rent = new forRent({
     type: req.body.type,
     bedrooms: req.body.bedrooms,
@@ -35,10 +35,20 @@ exports.postRent = async (req, res, next) => {
     state: req.body.state,
     city: req.body.city,
     neighbourhood: req.body.neighbourhood,
+    user: req.userId,
   });
   try {
     const postRent = await rent.save();
-    res.status(201).json({ message: "Post Saved", postRent: postRent });
+    const user = await User.findById(req.userId);
+    await user.items.push(rent);
+    await user.save();
+    // user.items.push(rent);
+    console.log(user);
+    res.status(201).json({
+      message: "Post Saved",
+      postRent: postRent,
+      userSave: user,
+    });
   } catch (error) {
     if (!error.statusCode) {
       error.statusCode = 500;
@@ -80,6 +90,7 @@ exports.updateRent = async (req, res, next) => {
     city: req.body.city,
     neighbourhood: req.body.neighbourhood,
   };
+
   if (req.files) {
     images = req.files;
   }
@@ -89,13 +100,19 @@ exports.updateRent = async (req, res, next) => {
     throw error;
   }
   try {
+    const check = await forRent.findById(postId);
+    if (check.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't update the Post");
+      error.statusCode = 403;
+      throw error;
+    }
     const rentPost = await forRent.findByIdAndUpdate(postId, payload);
     if (!rentPost) {
       const error = new Error("No Post Found");
       error.statusCode = 401;
       throw error;
     }
-    //CODE HERE TO CHECK USER THAT CREATE THE POST HAS THE RIGHT TO DELETE IT 
+    //CODE HERE TO CHECK USER THAT CREATE THE POST HAS THE RIGHT TO DELETE IT
     if (images !== rentPost.images) {
       rentPost.images.forEach((obj) => {
         const extract = obj.path.replace("\\", "/");
@@ -125,12 +142,25 @@ exports.deleteRent = async (req, res, next) => {
       error.statusCode = 401;
       throw error;
     }
+
     //CODE HERE TO CHECK THE USER WHICH CREATE THE POST HAS ONLY RIGHT TO DELETE IT
+
+    if (fetchPost.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't update the Post");
+      error.statusCode = 403;
+      throw error;
+    }
     fetchPost.images.forEach((obj) => {
       const extract = obj.path.replace("\\", "/");
       clearImage(extract);
     });
+  
+    
     await forRent.findByIdAndDelete(postId);
+    const user = await User.findById(req.userId);
+    user.items.pull(postId)
+    user.save();
+
     res.status(200).json({ message: "Post deleted successfully" });
   } catch (error) {
     if (!error.statusCode) {
@@ -139,6 +169,8 @@ exports.deleteRent = async (req, res, next) => {
     next(error);
   }
 };
+
+//forSale
 
 exports.postSale = async (req, res) => {
   console.log("Top");
@@ -166,9 +198,13 @@ exports.postSale = async (req, res) => {
     city: req.body.city,
     neighbour: req.body.neighbour,
     day: req.body.day,
+    user: req.userId,
   });
   try {
     const saved = await sale.save();
+    const user = await User.findById(req.userId);
+    await user.items.push(sale);
+    await user.save();
     res.status(201).json({ message: "Post Saved successfully", post: saved });
   } catch (error) {
     if (!error.statusCode) {
@@ -221,6 +257,12 @@ exports.updateSalePost = async (req, res, next) => {
     throw error;
   }
   try {
+    const check = await forSale.findById(postId);
+    if (check.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't update the Post");
+      error.statusCode = 403;
+      throw error;
+    }
     let salePost = await forSale.findByIdAndUpdate(postId, payload);
     if (!salePost) {
       const error = new Error("No Post Found");
@@ -258,6 +300,11 @@ exports.deleteSalePost = async (req, res, next) => {
       throw error;
     }
     //CODE HERE TO CHECK THE USER WHICH CREATE THE POST HAS ONLY RIGHT TO DELETE IT
+    if (findPost.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't Delete the Post");
+      error.statusCode = 403;
+      throw error;
+    }
 
     findPost.imageUrl.forEach((obj) => {
       const extract = obj.path.replace("\\", "/");
@@ -273,6 +320,8 @@ exports.deleteSalePost = async (req, res, next) => {
     next(error);
   }
 };
+
+// landsAndPlots
 
 exports.postPlots = async (req, res, next) => {
   // res.json({message:"postPlots"})
@@ -291,9 +340,13 @@ exports.postPlots = async (req, res, next) => {
     state: req.body.state,
     city: req.body.city,
     neighbourhood: req.body.neighbourhood,
+    user: req.userId,
   });
   try {
     const onGo = await newPost.save();
+    const user = await User.findById(req.userId);
+    await user.items.push(newPost);
+    await user.save();
     res.status(200).json({ message: "Post Saved Successfully", Post: onGo });
   } catch (error) {
     if (!error.statusCode) {
@@ -337,6 +390,12 @@ exports.updatePlot = async (req, res, next) => {
     throw error;
   }
   try {
+    const check = await landsAndPlots.findById(postId);
+    if (check.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't update the Post");
+      error.statusCode = 403;
+      throw error;
+    }
     const updtePlot = await landsAndPlots.findByIdAndUpdate(postId, payload);
     if (!updtePlot) {
       const error = new Error("No Post Found With this Id");
@@ -373,6 +432,12 @@ exports.deletePlot = async (req, res, next) => {
       throw error;
     }
     //CODE HERE TO CHECK USER AUTHORIZATION TO DELETE POST
+
+    if (fetchPlot.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't update the Post");
+      error.statusCode = 403;
+      throw error;
+    }
     fetchPlot.images.forEach((obj) => {
       const extract = obj.path.replace("\\", "/");
       clearImage(extract);
@@ -407,10 +472,16 @@ exports.postAdd = async (req, res, next) => {
     state: req.body.state,
     city: req.body.city,
     neighbourhood: req.body.neighbourhood,
+    user: req.userId,
   });
   try {
     const post = await postProd.save();
-    res.status(200).json({ message: "Post added successfully", post: post });
+    const user = await User.findById(req.userId);
+    await user.items.push(postProd);
+    await user.save();
+    res
+      .status(200)
+      .json({ message: "Post added successfully", post: post, user: user });
   } catch (error) {
     if (!error.statusCode) {
       error.statusCode = 500;
@@ -443,7 +514,6 @@ exports.UpdateAddShopOffice = async (req, res, next) => {
     city: req.body.city,
     neighbourhood: req.body.neighbourhood,
   };
-  //  CODE HERE TO CHECK AUTHORIZATION OF USER BY JWT
   if (req.files) {
     images = req.files;
   }
@@ -453,6 +523,12 @@ exports.UpdateAddShopOffice = async (req, res, next) => {
     throw error;
   }
   try {
+    const check = await forRentOfficeShops.findById(postId);
+    if (check.user.toString() !== req.userId) {
+      const error = new Error("Current Login user Can't update the Post");
+      error.statusCode = 403;
+      throw error;
+    }
     const updateAdd = await forRentOfficeShops.findByIdAndUpdate(
       prodId,
       payload
